Removed MainService. The servers and remote manager are now held by MainCoordinator, which is responsible for managing them. Both servers now share an eventloop, which is also owned by MainCoordinator.
