+++ TODO.rtf
@@ -128,7 +128,9 @@ Remote:\
 - Ensure all Remotes are sharing a single eventloop group\
 - \cf9 Implement transfer speed calculation\cf3  \
 \
-Network: \cf0 \
+Network: \
+	- Android app is still occasionally failing because iOS won\'92t fucking let me stop broadcasting fucking IPv6 because jUsT LEt tHe FrAmEWorK DeCidE \
+	WhIcH InTerFaCE tO uSe\cf0 \
 	- Investigate/Fix slow connection bullshit\
 	- Handle network change \
 		- Servers need to restart on new ip addresses\
